fix: wip fix for save and go back
